fix safe lock rotation comparison

+++ Unity project/Assets/Scripts/UseableItems/Actions/TurnLock.cs
@@ -3,13 +3,13 @@ using System.Collections;
 
 public class TurnLock : RotateOnUse
 {
-	public delegate void OnTurnLock(float rot);
+	public delegate void OnTurnLock(short rot);
 	public static event OnTurnLock LockTurned;
-	private float _degrees = 0;
+	private short _degrees = 0;
 	override protected void doUse()
 	{
 		base.doUse ();
-		_degrees += Rotation.y;
+		_degrees += (short)Rotation.y;
 		LockTurned(_degrees);
 	}
 }
